refactor: 경로명 수정
