+++ dao/project.js
@@ -1,5 +1,5 @@
 const pool = require('./db');
-const { customError } = require('../utils/errors/customError');
+const { customError } = require('../utils/errors/custom');
 
 const createProject = async (data) => {
   const conn = await pool.getConnection();
