+++ dao/user.js
@@ -3,7 +3,7 @@ const admin = require('firebase-admin');
 
 const pool = require('./db');
 const { firebaseError } = require('../utils/errors/firebase');
-const { customError } = require('../utils/errors/customError');
+const { customError } = require('../utils/errors/custom');
 
 const checkNickname = async (nickname) => {
   const conn = await pool.getConnection();
