+++ services/notice.js
@@ -1,7 +1,7 @@
 const noticeDao = require('../dao/notice');
 
 const { toBoolean, cutId } = require('../utils/query');
-const { customError } = require('../utils/errors/customError');
+const { customError } = require('../utils/errors/custom');
 
 const createNotice = async ({ study_id }, createData) => {
   createData.study_id = study_id;
